Tidy up and document code
Made a start on this. More to come in later commits.
