+++ internal/embed/gen_bindata.go
@@ -189,6 +189,13 @@ import "list"
 #CommandFile: #Guide.#CommandFile
 #Upload:      #Guide.#Upload
 #UploadFile:  #Guide.#UploadFile
+
+#GenConfig: [string]: #PrestepConfig
+
+#PrestepConfig: {
+	Endpoint: string
+	Networks: [...string]
+}
 `)
 
 func preguide_cue() ([]byte, error) {
