+++ testdata/env_template.txt
@@ -22,14 +22,14 @@ preguide gen -config conf.cue -out _output
 ! stdout .+
 ! stderr .+
 cmp _output/_posts/env_template.markdown env_template/en.markdown.golden
-cmp env_template/en_testlog.txt env_template/en_testlog.txt.golden
+cmp env_template/en_log.txt env_template/en_log.txt.golden
 
 # Run with -raw
 preguide gen -raw -config conf.cue -out _output
 cmp stdout env_template/raw.cue.golden
 ! stderr .+
 cmp _output/_posts/env_template.markdown env_template/en.markdown.raw.golden
-cmp env_template/en_testlog.txt env_template/en_testlog.txt.raw.golden
+cmp env_template/en_log.txt env_template/en_log.txt.raw.golden
 
 -- prestep.txt --
 {
@@ -76,7 +76,7 @@ title: A test with all directives
 $ echo -n "The answer is: {{.GREETING}}!"
 The answer is: {{.GREETING}}!
 ```
--- env_template/en_testlog.txt.golden --
+-- env_template/en_log.txt.golden --
 Terminals: [
   {
     "Name": "term1",
@@ -137,7 +137,7 @@ title: A test with all directives
 $ echo -n "The answer is: Hello, world!!"
 The answer is: Hello, world!!
 ```
--- env_template/en_testlog.txt.raw.golden --
+-- env_template/en_log.txt.raw.golden --
 Terminals: [
   {
     "Name": "term1",
