+++ util.go
@@ -2,56 +2,69 @@ package main
 
 import (
 	"fmt"
-	"strconv"
+	"regexp"
 	"strings"
 )
 
-// Copyright 2011 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// split breaks the line into words
-func split(line string) ([]string, error) {
-	// Parse line, obeying quoted strings.
-	var words []string
-	// One (possibly quoted) word per iteration.
-Words:
-	for {
-		line = strings.TrimLeft(line, " \t")
-		if len(line) == 0 {
-			break
+// check verifies that err is nil, else it parnics wrapping err in a knownErr
+// (which is recovered by my mainerr). This allows clean, fluent code without
+// lots of error handling, where that error handling would otherwise simply
+// bubble an error up the stack.
+func check(err error, format string, args ...interface{}) {
+	if err != nil {
+		if format != "" {
+			err = fmt.Errorf(format, args...)
 		}
-		if line[0] == '"' {
-			for i := 1; i < len(line); i++ {
-				c := line[i] // Only looking for ASCII so this is OK.
-				switch c {
-				case '\\':
-					if i+1 == len(line) {
-						return nil, fmt.Errorf("bad backslash")
-					}
-					i++ // Absorb next byte (If it's a multibyte we'll get an error in Unquote).
-				case '"':
-					word, err := strconv.Unquote(line[0 : i+1])
-					if err != nil {
-						return nil, fmt.Errorf("bad quoted string")
-					}
-					words = append(words, word)
-					line = line[i+1:]
-					// Check the next character is space or end of line.
-					if len(line) > 0 && line[0] != ' ' && line[0] != '\t' {
-						return nil, fmt.Errorf("expect space after quoted argument")
-					}
-					continue Words
-				}
-			}
-			return nil, fmt.Errorf("mismatched quoted string")
-		}
-		i := strings.IndexAny(line, " \t")
-		if i < 0 {
-			i = len(line)
-		}
-		words = append(words, line[0:i])
-		line = line[i:]
+		panic(knownErr{err})
 	}
-	return words, nil
+}
+
+// raise raises a knownErr, wrapping a fmt.Errorf generated error using the
+// provided format and args. See the documentation for check on why these
+// functions exist.
+func raise(format string, args ...interface{}) {
+	panic(knownErr{fmt.Errorf(format, args...)})
+}
+
+// knownErr is the sentinel error type used by check and raise. Values of this
+// type are recovered in mainerr. See thd documentation for check for more
+// details.
+type knownErr struct{ error }
+
+// handleKnown is a convenience function used in a defer to recover from a
+// knownErr. See the usage in mainerr.
+func handleKnown(err *error) {
+	switch r := recover().(type) {
+	case nil:
+	case knownErr:
+		*err = r
+	default:
+		panic(r)
+	}
+}
+
+// stringFlagList is a supporting type for generating a string flag that can
+// appear multiple times.
+type stringFlagList struct {
+	vals *[]string
+}
+
+func (s stringFlagList) String() string {
+	if s.vals == nil {
+		return ""
+	}
+	return strings.Join(*s.vals, " ")
+}
+
+func (s stringFlagList) Set(v string) error {
+	*s.vals = append(*s.vals, v)
+	return nil
+}
+
+var markdownFile = regexp.MustCompile(`.(md|mkdn?|mdown|markdown)$`)
+
+// isMarkdown determines whether name is a markdown file name
+func isMarkdown(name string) (string, bool) {
+	ext := markdownFile.FindString(name)
+	return ext, ext != ""
 }
